# ENH: stats.pareto fit improvement for parameter combinations