--- a/scipy/stats/_continuous_distns.py
+++ b/scipy/stats/_continuous_distns.py
@@ -6123,18 +6123,23 @@ class pareto_gen(rv_continuous):
         parameters = _check_fit_input_parameters(self, data, args, kwds)
         data, fshape, floc, fscale = parameters
 
+        # ensure that any fixed parameters don't violate constraints of the
+        # distribution before continuing.
+        if floc is not None and np.min(data) - floc < (fscale or 0):
+            raise FitDataError("pareto", lower=1, upper=np.inf)
+
         ndata = data.shape[0]
 
         def get_shape(scale, location):
-            # The partial derivative w.r.t. the shape can be solved in
-            # closed form for the shape.
+            # The first-order necessary condition on `shape` can be solved in
+            # closed form
             return ndata / np.sum(np.log((data - location) / scale))
 
         if floc is fscale is None:
-            # Optimization over a system of equations is possible when the
-            # location of the data is not set. The optimization requires that
-            # the scale be free, but the shape may be fixed. The equations
-            # are from the partial derivatives of the log-likelihood function.
+            # The support of the distribution is (x - loc)/scale > 0
+            # The method of Lagrange multipliers turns this constraint
+            # into an equation that can be solved numerically.
+            # See gh-12545 for details.
 
             def dLdScale(shape, scale):
                 # The partial derivative of the log-likelihood function w.r.t.
@@ -6147,8 +6152,8 @@ class pareto_gen(rv_continuous):
                 return (shape + 1) * np.sum(1 / (data - location))
 
             def fun_to_solve(scale):
-                # optimize over the scale, setting the partial derivatives
-                # w.r.t. to location and scale equal.
+                # optimize the scale by setting the partial derivatives
+                # w.r.t. to location and scale equal and solving.
                 location = np.min(data) - scale
                 shape = fshape or get_shape(scale, location)
                 return dLdLocation(shape, location) - dLdScale(shape, scale)
@@ -6164,15 +6169,16 @@ class pareto_gen(rv_continuous):
             # if a root is not between the brackets, iteratively expand them
             # until they include a sign change, checking after each bracket is
             # modified.
-            while not interval_contains_root(lbrack, rbrack):
+            while (not interval_contains_root(lbrack, rbrack)
+                   and (lbrack > 0 or rbrack < np.inf)):
                 lbrack /= 2
                 rbrack *= 2
             res = root_scalar(fun_to_solve, bracket=[lbrack, rbrack])
 
             if res.converged:
-                fscale = res.root
-                floc = np.min(data) - fscale
-                fshape = fshape or get_shape(fscale, floc)
+                scale = res.root
+                loc = np.min(data) - scale
+                shape = fshape or get_shape(scale, loc)
 
                 # The Pareto distribution requires that its parameters satisfy
                 # the condition `fscale + floc <= min(data)`. However, to
@@ -6180,17 +6186,14 @@ class pareto_gen(rv_continuous):
                 # is strictly less than `min(data)`. If this condition
                 # is not satisfied, reduce the scale with `np.nextafter` to
                 # ensure that data does not fall outside of the support.
-                if not (fscale + floc) < np.min(data):
-                    fscale = np.min(data) - floc
-                    fscale = np.nextafter(fscale, 0)
-                return fshape, floc, fscale
+                if not (scale + loc) < np.min(data):
+                    scale = np.min(data) - floc
+                    scale = np.nextafter(scale, 0)
+                return shape, loc, scale
             else:
                 return super().fit(data, **kwds)
         elif floc is None:
-            floc = np.min(data) - fscale
-
-        if np.any(data - floc < (fscale if fscale else 0)):
-            raise FitDataError("pareto", lower=1, upper=np.inf)
+            loc = np.min(data) - fscale
 
         # Source: Evans, Hastings, and Peacock (2000), Statistical
         # Distributions, 3rd. Ed., John Wiley and Sons. Page 149.
@@ -6199,6 +6202,7 @@ class pareto_gen(rv_continuous):
         if fshape is None:
             fshape = get_shape(fscale, floc)
         return fshape, floc, fscale
+
 
 
 pareto = pareto_gen(a=1.0, name="pareto")