--- a/scipy/stats/_continuous_distns.py
+++ b/scipy/stats/_continuous_distns.py
@@ -6297,7 +6297,7 @@ class pareto_gen(rv_continuous):
 
             # set brackets for `root_scalar` to use when optimizing over the
             # scale such that a root is likely between them. Use user supplied
-            # guess or default 1. 
+            # guess or default 1.
             brack_start = kwds.get('scale', 1)
             lbrack, rbrack = brack_start / 2, brack_start * 2
             # if a root is not between the brackets, iteratively expand them