--- a/scipy/stats/tests/test_distributions.py
+++ b/scipy/stats/tests/test_distributions.py
@@ -5242,7 +5242,8 @@ class TestStudentizedRange:
 
     @pytest.mark.slow
     def test_moment_vectorization(self):
-        # Test moment broadcasting. Calls `_munp` directly         # `rv_continuous.moment` is broken at time of writing. See gh-12192
+        # Test moment broadcasting. Calls `_munp` directly
+        # `rv_continuous.moment` is broken at time of writing. See gh-12192
         m = stats.studentized_range._munp([1, 2], [4, 5], [10, 11])
         assert_allclose(m.shape, (2,))