#### What does this implement/fix?
<!--Please explain your changes.-->

This PR improves fit speed and fit accuracy for the `stats.pareto` distribution. This allows improved fitting for all combinations of fixed and free parameters. All combinations are tested in the unit test suite to ensure their accuracy is better than `rv_continuous.fit` (measured by evaluation of the penalized log-likelihood function). 

All combinations are shown to be quicker with ASV benchmarks. (The ASV webpage was misbehaving, so matplotlib will have to do) This shows the improvement of fitting when `floc` is free, which is what this PR addresses. Analytical fits for when `floc` is not fixed are already available.


![image](https://user-images.githubusercontent.com/44255917/153311712-d9221e38-ecfa-4ecf-9cf2-ea2916dc2494.png)
Raw output of ASV runs, [script for above plot](https://gist.github.com/4f94b31e7123293007fb540459455572):
<details>

<pre>
## BEFORE
[100.00%] ··· ============== =========== ============= ==================== ===================== ===================== ====================== ===================== ====================== ====================== =======================
              --                                                                                                            shape1_fixed / shape2_fixed / shape3_fixed
              ---------------------------------------- -----------------------------------------------------------------------------------------------------------------------------------------------------------------------------------
               distribution   loc_fixed   scale_fixed   True / True / True   True / True / False   True / False / True   True / False / False   False / True / True   False / True / False   False / False / True   False / False / False
              ============== =========== ============= ==================== ===================== ===================== ====================== ===================== ====================== ====================== =======================
                  pareto         True         True             n/a                   n/a                   n/a                   n/a                    n/a                   n/a                    n/a                   39.3±3μs
                  pareto         True        False             n/a                   n/a                   n/a                 29.6±3μs                 n/a                   n/a                    n/a                   45.1±2μs
                  pareto        False         True             n/a                   n/a                   n/a                 16.5±4ms                 n/a                   n/a                    n/a                   28.5±2ms
                  pareto        False        False             n/a                   n/a                   n/a                 49.8±2ms                 n/a                   n/a                    n/a                   35.6±9ms
              ============== =========== ============= ==================== ===================== ===================== ====================== ===================== ====================== ====================== =======================


### AFTER



[100.00%] ··· ============== =========== ============= ==================== ===================== ===================== ====================== ===================== ====================== ====================== =======================
              --                                                                                                            shape1_fixed / shape2_fixed / shape3_fixed
              ---------------------------------------- -----------------------------------------------------------------------------------------------------------------------------------------------------------------------------------
               distribution   loc_fixed   scale_fixed   True / True / True   True / True / False   True / False / True   True / False / False   False / True / True   False / True / False   False / False / True   False / False / False
              ============== =========== ============= ==================== ===================== ===================== ====================== ===================== ====================== ====================== =======================
                  pareto         True         True             n/a                   n/a                   n/a                   n/a                    n/a                   n/a                    n/a                  35.7±0.9μs
                  pareto         True        False             n/a                   n/a                   n/a                21.4±0.2μs                n/a                   n/a                    n/a                   42.5±2μs
                  pareto        False         True             n/a                   n/a                   n/a                16.8±0.2μs                n/a                   n/a                    n/a                  36.4±0.7μs
                  pareto        False        False             n/a                   n/a                   n/a                 878±40μs                 n/a                   n/a                    n/a                  2.05±0.1ms
              ============== =========== ============= ==================== ===================== ===================== ====================== ===================== ====================== ====================== =======================



</pre>
</details>

The math behind this implementation is briefly described in in-line comments, and a full discussion can be found here: https://github.com/scipy/scipy/issues/12545#issuecomment-735313963. Optional reading is additional discussion found here: https://github.com/mdhaber/scipy/pull/68#event-6037231133

#### Additional information
<!--Any additional information you think is important.-->
We could consider closing this issue: https://github.com/scipy/scipy/issues/12545 because it addresses runtimewarnings originating from `rv_continuous.fit`, which is only used here as a last resort. It is still possible to get runtimewarnings with this method with extreme data, but it is infrequent, whereas with `rv_continous.fit` it happened every time when all parameters were free.