@mdhaber 2022-02-23
> It is still possible to get runtimewarnings with this method with extreme data,

Still? In what circumstances?