@mdhaber 2022-03-02
We've run into two main issue with `root_scalar` bracket methods here:

- Our function to solve asymptotically approaches zero at `scale = np.inf`. When there is a true zero crossing within the bracket but the right end of the bracket is too large, `root_scalar` thinks the right end of the bracket is the root. I can't really blame `root_scalar` for this; it just needs to be given a better bracket. We solved this for the majority of cases by starting our search for a bracket centered around `scale = 1` instead of `scale = np.ptp(data)`. I think that is OK for now. If this ever comes up as an issue, we can recommend that the user provide a guess for `scale`. 
- There seem to be pathological cases where there is no true zero-crossing and the asymptote at `scale = np.inf` (or some super high value after precision is lost) really is the correct answer. In the case above, however, `root_scalar` did not converge to the right end of the bracket; instead, it converged to `scale = 0` (approximately) and reported success even though the value of the function was nowhere close to zero. This is surprising because the right end of the bracket has a much lower value of the function to be solved. To be fair, `root_scalar` could have been thrown off because the value of the function at the left end of the bracket was infinite. Your proposed solution above was to ensure that the function value at the ends of the bracket is finite. I hesitate to implement this because it seems that `root_scalar` could have done better in this case. Rather than adding this extra logic to protect against this pathological case here, I'd suggest leaving the logic here as-is and opening an issue against `root_scalar` with a M.W.E. of this problem. As an M.W.E., I thought this would work:

```python3
def f(x):
    return 1/x - 1e-10
from scipy.optimize import root_scalar
root_scalar(f, bracket=(0, 1e13))
```
but it doesn't. `root_scalar` correctly finds the root. So I still think we should leave the logic as-is for now, but we need to find out why `root_scalar` seems to be doing the wrong thing here and see if we can find other examples.