@swallan 2022-03-05
> This is surprising because the right end of the bracket has a lower value of the function to be solved (correct?).

correct

----
I tried also fitting this situation with `root_scalar` and `method="newton"`, and the result was a failure to fit. 

Since I did this in the debugger this isn't runnable, but here's what happens:
```
IPdb [23]: from scipy.optimize._numdiff import approx_derivative

IPdb [24]: def fp(x):
      ...:     return approx_fprime((x,),fun_to_solve)[0]
      ...:     

IPdb [25]: root_scalar(fun_to_solve, method='newton', x0=1, fprime=fp)
      converged: False
           flag: 'convergence error'
 function_calls: 40
     iterations: 20
           root: 348299.76238276693
/Users/swallan/Documents/SciPy/scipy_meson/scipy/build-install/lib/python3.10/site-packages/scipy/optimize/_zeros_py.py:302: RuntimeWarning: Derivative was zero.
  warnings.warn(msg, RuntimeWarning)
```
It says the derivative is zero, which is true:
```
fp(348299.76238276693)
0.0
```
So this really might be a pathological case for which there is not a solution. In any case, `root_scalar` isn't behaving like we'd expect. 

> see if we can find a simple example to open an issue about.

I can open a follow up issue about that.