@swallan 2022-03-07
Relating to the pathological `levy_l` case, I also tried solving `root_scalar` with `method="newton"` by providing `fprime`( `scipy.optimize.approx_fprime` of `fun_to_solve`) and it fails to converge to the root due to the derivative being zero:

```
IPdb [6]: root_scalar(fun_to_solve, method='newton', x0=1000, fprime=fp)
      converged: False
           flag: 'convergence error'
 function_calls: 22
     iterations: 11
           root: 314954.1208635712
/Users/swallan/Documents/SciPy/scipy_meson/scipy/build-install/lib/python3.10/site-packages/scipy/optimize/_zeros_py.py:302: RuntimeWarning: Derivative was zero.
  warnings.warn(msg, RuntimeWarning)
```

so there may not actually be a true root. This is quite a bit more friendly of a result than the bracketing method which claimed to converge.

Newton is quite a bit slower for data that originates from the pareto distribution, and seems to only work well if the initial guess is good. So keeping the bracket method seems like the best option.