@tupui 2022-03-09
> @tupui Did you have other thoughts? I think your comments have been addressed. There are a couple comments but few / no changes originated from [that discussion](https://github.com/scipy/scipy/pull/15567#issuecomment-1057420611).

I did not follow all the discussions. Matt seems to be on it so I will let him continue unless you want a final pass at the end.

> I think the test failure is a timeout (it cancelled after 60 minutes)

Yes no problem, happens a lot on Azure...